Minor reorganisation and DRY fixes.
