+++ pdr_backend/pdr_dashboard/callbacks/callbacks_home.py
@@ -5,11 +5,11 @@ from dash import Input, Output, State
 from pdr_backend.cli.arg_feeds import ArgFeeds
 from pdr_backend.pdr_dashboard.dash_components.plots import get_figures_and_metrics
 from pdr_backend.pdr_dashboard.dash_components.view_elements import get_graph
-from pdr_backend.pdr_dashboard.util.data import (
+from pdr_backend.pdr_dashboard.util.format import format_value
+from pdr_backend.pdr_dashboard.util.helpers import (
     get_date_period_text_for_selected_predictoors,
     select_or_clear_all_by_table,
 )
-from pdr_backend.pdr_dashboard.util.format import format_value
 
 
 # pylint: disable=too-many-statements
